move tz and eth method tools to modules
